RateDoc Tests/ratedoc-tests/e2e/requirements.txt
pytest-playwright
playwright
pytest-base-url
pytest-html
